+++ AppFitness/ViewControllerCalendar.m
@@ -61,7 +61,7 @@
         
         NSDateFormatter *dt = [[NSDateFormatter alloc] init];
         NSString *strMonthName = [[dt monthSymbols] objectAtIndex:month-1];//January,Febryary,March etc...
-        strMonthName = [strMonthName stringByAppendingString:[NSString  stringWithFormat:@"- %d",year]];
+        strMonthName = [strMonthName stringByAppendingString:[NSString  stringWithFormat:@"- %ld",(long)year]];
         
         X = isLeft ? 10 : 164;
         
@@ -82,7 +82,7 @@
         vwCal.layer.borderColor = [UIColor blackColor].CGColor;
         vwCal.layer.borderWidth = 2.0;
         
-        vwCal = [vwCal createCalOfDay:day Month:month Year:year MonthName:strMonthName];
+        vwCal = [vwCal createCalOfDay:(int)day Month:(int)month Year:(int)year MonthName:strMonthName];
         [viewTmp addSubview:vwCal];
         isLeft = !isLeft;
     }
